cosmetic refactor of PiPCA implementation
